Fix failing test case where `uri.host` was null.
Font names are passed in without a host name or protocol, so
trying to use `uri.host` was resulting in "null" being inserted
into the path. This fixes that by checking whether `uri.host` is
falsey before using it.

Also, it was getting hard to read, so factored out some
expressions into their own variables.